adamg codereview
