+++ soloact/__init__.py
@@ -1,0 +1,3 @@
+from soloact.make_paths import make_source_paths
+
+from soloact.data import *
