+++ soloact/data/__init__.py
@@ -1,0 +1,2 @@
+from soloact.data.annotations_chords import write_annotation, write_chords
+from soloact.data.make_dataset import augment_data
